lambda-functions/upscaler/requirements.txt
Pillow==10.1.0
boto3==1.34.0